we no longer even try to handle section editing
Section editing is only available to users who can write the page. User
who can write the page have access to all data in it, so hiding content
from them with ifauthex is useless in the first place.

For proper handling, we might need to overwrite the xhtml renderer and
tune a lot of internal variables and properties.